Refactor of all the admin pages so it is more organised
